Add favicons and logo
The favicons double as a visual indicator on tabs to show whether a page is in
view mode or edit mode. This is very useful to me because I often have
multiple pages open in separate tabs and this will show at a glance which
ones I'm in the middle of editing.