Include instructions and script for data dump/restore
We need a way to export and import data into the database, for a variety
of reasons but mostly because we might need to migrate data to a new
version of the schema from time to time. (Which we hope will be rare.)
And developing a whole database migration system for this little project,
while good practice, is not worth the effort just yet.