Fix SQLite full-text search index
Unforuntately, this change requires a database migration. See the "Data Export
and Import" section in `README.md` for instructions on how to do this.

It turns out we can't (and never should have) used an FTS5 index linked to an
external content table. Every row in an index linked to an external table must
always be kept precisely in sync with every row in the external table. (I'm not
sure why they designed it that way, but there must be a good reason.) In
particular, you cannot simply choose to index only certain rows. Removing rows
from either the content table or the index results in a corrupted index. In
Silicon, this manifested as results that did not contain a term in the search
query, or outright errors and exceptions being thrown.

The fix is to use a "normal" FTS5 index, which is an FTS5 index mode that
doesn't seem to have a name and is not described well in the docs or in
articles/tutorials which attempt to teach FTS5. In contrast to external content
or contentless tables, there are few restrictions on what you can do with a
normal index. You can add or remove whatever data you like using normal SQL
statements. The trade-off is that it apparently takes up more space.

We also drop the DELETE trigger because page/revision deletion is not
implemented yet anyway, and there are corner cases that need additional
consideration.