Refactor colors to use a ‘color token’ system
The purpose of this change is to make color theming easier, dark mode in particular. It's based on my experience working at a Big Tech Company that went through a couple of design system strategies. I do this for my personal projects and find it makes things easier even for a 1-person "team."

The basic idea is to split your color system into 3 layers:

- Palette, the set of all colors
- Color tokens, mapping colors to generic and reusable names with semantic meaning
- Component use, assigning a color token to a specific place in the UI

A hypothetical dark mode theme would remap only the color tokens layer.