Don't use prettier on Markdown files
Prettier does "opinionated" (read: annoying) things to unordered lists in
Markdown. See comments in file.

I would be far less angry about this if I could apply the fix in `.prettierrc`
instead of adding a whole new file to the top of the tree. If I run across
another prettier issue that ends in me writing a salty commit message, I'm
removing it from this project.