exporter.py: slugify page titles in JSON and file names
It _should_ be impossible for non-slugified page titles to ened up in the
database, but I found a few in my own collection. I can't seem to reproduce
creating pages with these titles so my guess is that they came from an earlier
verison of Silicon before slugification was as strict as it is now.